# swagger.yml
swagger: '2.0'
info:
  title: forum
  description: |
    Тестовое задание для реализации проекта "Форумы" на курсе по базам данных в
    Технопарке VK (https://park.vk.company).
  version: "0.1.0"
schemes:
  - http
  - https
basePath: /api
consumes:
  - application/json
produces:
  - application/json
paths:
  /forum/create:
    post:
      summary: Создание форума
      tags:
        - forum
      description: |
        Создание нового форума.
      operationId: forumCreate
      parameters:
        - name: forum
          in: body
          description: Данные форума.
          required: true
          schema:
            $ref: '#/components/schemas/Forum'
      responses:
        201:
          description: |
            Форум успешно создан.
            Возвращает данные созданного форума.
          schema:
            $ref: '#/components/schemas/Forum'
        404:
          description: |
            Владелец форума не найден.
          schema:
            $ref: '#/components/schemas/Error'
        409:
          description: |
            Форум уже присутсвует в базе данных.
            Возвращает данные ранее созданного форума.
          schema:
            $ref: '#/components/schemas/Forum'
  /forum/{slug}/details:
    get:
      summary: Получение информации о форуме
      tags:
        - forum
      description: |
        Получение информации о форуме по его идентификаторе.
      consumes: [ ]
      operationId: forumGetOne
      parameters:
        - name: slug
          in: path
          description: Идентификатор форума.
          required: true
          type: string
          format: identity
          schema:
            $ref: '#/components/schemas/Slug'
      responses:
        200:
          description: |
            Информация о форуме.
          schema:
            $ref: '#/components/schemas/Forum'
        404:
          description: |
            Форум отсутсвует в системе.
          schema:
            $ref: '#/components/schemas/Error'
  /forum/{slug}/create:
    post:
      summary: Создание ветки
      tags:
        - forum
      description: |
        Добавление новой ветки обсуждения на форум.
      operationId: threadCreate
      
      parameters:
        - name: slug
          in: path
          description: Идентификатор форума.
          required: true
          type: string
          format: identity
          schema:
            $ref: '#/components/schemas/Slug'
      responses:
        201:
          description: |
            Ветка обсуждения успешно создана.
            Возвращает данные созданной ветки обсуждения.
          schema:
            $ref: '#/components/schemas/Thread'
        404:
          description: |
            Автор ветки или форум не найдены.
          schema:
            $ref: '#/components/schemas/Error'
        409:
          description: |
            Ветка обсуждения уже присутсвует в базе данных.
            Возвращает данные ранее созданной ветки обсуждения.
          schema:
            $ref: '#/components/schemas/Thread'
  /forum/{slug}/users:
    get:
      summary: Пользователи данного форума
      tags:
        - forum
      description: |
        Получение списка пользователей, у которых есть пост или ветка обсуждения в данном форуме.

        Пользователи выводятся отсортированные по nickname в порядке возрастания.
        Порядок сотрировки должен соответсвовать побайтовому сравнение в нижнем регистре.
      consumes: [ ]
      operationId: forumGetUsers
      parameters:
        - name: slug
          in: path
          description: Идентификатор форума.
          required: true
          type: string
          format: identity
          schema:
            $ref: '#/components/schemas/Slug'
        - name: limit
          in: query
          type: number
          
          default: 100
          minimum: 1
          maximum: 10000
          description: Максимальное кол-во возвращаемых записей.
          schema:
            $ref: '#/components/schemas/Limit'
        - name: since
          in: query
          type: string
          format: identity
          description: |
            Идентификатор пользователя, с которого будут выводиться пользоватли
            (пользователь с данным идентификатором в результат не попадает).
          schema:
            $ref: '#/components/schemas/Since'
        - name: desc
          in: query
          type: boolean
          description: |
            Флаг сортировки по убыванию.
          schema:
            $ref: '#/components/schemas/Desc'
      responses:
        200:
          description: |
            Информация о пользователях форума.
          schema:
            $ref: '#/components/schemas/Users'
        404:
          description: |
            Форум отсутсвует в системе.
          schema:
            $ref: '#/components/schemas/Error'
  /forum/{slug}/threads:
    get:
      summary: Список ветвей обсужления форума
      tags:
        - forum
      description: |
        Получение списка ветвей обсужления данного форума.

        Ветви обсуждения выводятся отсортированные по дате создания.
      consumes: [ ]
      operationId: forumGetThreads
      parameters:
        - name: slug
          in: path
          description: Идентификатор форума.
          required: true
          type: string
          format: identity
          schema:
            $ref: '#/components/schemas/Slug'
        - name: limit
          in: query
          type: number
          
          default: 100
          minimum: 1
          maximum: 10000
          description: Максимальное кол-во возвращаемых записей.
          schema:
            $ref: '#/components/schemas/Limit'
        - name: since
          in: query
          type: string
          format: date-time
          description: |
            Дата создания ветви обсуждения, с которой будут выводиться записи
            (ветвь обсуждения с указанной датой попадает в результат выборки).
          schema:
            $ref: '#/components/schemas/Since'
        - name: desc
          in: query
          type: boolean
          description: |
            Флаг сортировки по убыванию.
          schema:
            $ref: '#/components/schemas/Desc'
      responses:
        200:
          description: |
            Информация о ветках обсуждения на форуме.
          schema:
            $ref: '#/components/schemas/Threads'
        404:
          description: |
            Форум отсутсвует в системе.
          schema:
            $ref: '#/components/schemas/Error'
  /post/{id}/details:
    get:
      summary: Получение информации о ветке обсуждения
      tags:
        - post
      description: |
        Получение информации о ветке обсуждения по его имени.
      consumes: [ ]
      operationId: postGetOne
      parameters:
        - name: id
          in: path
          description: Идентификатор сообщения.
          required: true
          type: number
          
          schema:
            $ref: '#/components/schemas/Id'
        - name: related
          in: query
          type: array
          description: |
            Включение полной информации о соответвующем объекте сообщения.

            Если тип объекта не указан, то полная информация об этих объектах не
            передаётся.
          items:
            type: string
            enum:
              - user
              - forum
              - thread
          schema:
            $ref: '#/components/schemas/Related'
      responses:
        200:
          description: |
            Информация о ветке обсуждения.
          schema:
            $ref: '#/components/schemas/PostFull'
        404:
          description: |
            Ветка обсуждения отсутсвует в форуме.
          schema:
            $ref: '#/components/schemas/Error'
    post:
      summary: Изменение сообщения
      description: |
        Изменение сообщения на форуме.

        Если сообщение поменяло текст, то оно должно получить отметку `isEdited`.
      operationId: postUpdate
      parameters:
        - name: id
          in: path
          description: Идентификатор сообщения.
          required: true
          type: number
          
          schema:
            $ref: '#/components/schemas/Id'
        - name: post
          in: body
          description: Изменения сообщения.
          required: true
          schema:
            $ref: '#/components/schemas/PostUpdate'
      responses:
        200:
          description: |
            Информация о сообщении.
          schema:
            $ref: '#/components/schemas/Post'
        404:
          description: |
            Сообщение отсутсвует в форуме.
          schema:
            $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          readOnly: true
          description: |
            Текстовое описание ошибки.
            В процессе проверки API никаких проверок на содерижимое данного описание не делается.
          example: |
            Can't find user with id #42
...
    User:
      description: |
        Информация о пользователе.
      type: object
      properties:
        nickname:
          type: string
          format: identity
          readOnly: true
          description: |
            Имя пользователя (уникальное поле).
            Данное поле допускает только латиницу, цифры и знак подчеркивания.
            Сравнение имени регистронезависимо.
          example: j.sparrow
        fullname:
          type: string
          description: Полное имя пользователя.
          example: Captain Jack Sparrow
          x-isnullable: false
        about:
          type: string
          format: text
          description: Описание пользователя.
          example: This is the day you will always remember as the day that you almost caught Captain Jack Sparrow!
        email:
          type: string
          format: email
          description: Почтовый адрес пользователя (уникальное поле).
          example: captaina@blackpearl.sea
          x-isnullable: false
      required:
        - fullname
        - email
    Users:
      type: array
      items:
        $ref: '#/components/schemas/User'
...
    Forum:
      description: |
        Информация о форуме.
      type: object
      properties:
        title:
          type: string
          description: Название форума.
          example: Pirate stories
          x-isnullable: false
        user:
          type: string
          format: identity
          description: Nickname пользователя, который отвечает за форум.
          example: j.sparrow
          x-isnullable: false
        slug:
          type: string
          format: identity
          description: Человекопонятный URL (https://ru.wikipedia.org/wiki/%D0%A1%D0%B5%D0%BC%D0%B0%D0%BD%D1%82%D0%B8%D1%87%D0%B5%D1%81%D0%BA%D0%B8%D0%B9_URL), уникальное поле.
          pattern: ^(\d|\w|-|_)*(\w|-|_)(\d|\w|-|_)*$
          example: pirate-stories
          x-isnullable: false
        posts:
          type: number
          readOnly: true
          description: |
            Общее кол-во сообщений в данном форуме.
          example: 200000
        threads:
          type: number
          
          readOnly: true
          description: |
            Общее кол-во ветвей обсуждения в данном форуме.
          example: 200
      required:
        - title
        - user
        - slug
    Thread:
      description: |
        Ветка обсуждения на форуме.
      type: object
      properties:
        id:
          type: number
          
          description: Идентификатор ветки обсуждения.
          readOnly: true
          example: 42
        title:
          type: string
          description: Заголовок ветки обсуждения.
          example: Davy Jones cache
          x-isnullable: false
        author:
          type: string
          format: identity
          description: Пользователь, создавший данную тему.
          example: j.sparrow
          x-isnullable: false
        forum:
          type: string
          format: identity
          description: Форум, в котором расположена данная ветка обсуждения.
          readOnly: true
          example: pirate-stories
        message:
          type: string
          format: text
          description: Описание ветки обсуждения.
          example: An urgent need to reveal the hiding place of Davy Jones. Who is willing to help in this matter?
          x-isnullable: false
        votes:
          type: number
          
          description: Кол-во голосов непосредственно за данное сообщение форума.
          readOnly: true
        slug:
          type: string
          format: identity
          description: |
            Человекопонятный URL (https://ru.wikipedia.org/wiki/%D0%A1%D0%B5%D0%BC%D0%B0%D0%BD%D1%82%D0%B8%D1%87%D0%B5%D1%81%D0%BA%D0%B8%D0%B9_URL).
            В данной структуре slug опционален и не может быть числом.
          pattern: ^(\d|\w|-|_)*(\w|-|_)(\d|\w|-|_)*$
          readOnly: true
          example: jones-cache
        created:
          type: string
          format: date-time
          description: Дата создания ветки на форуме.
          example: 2017-01-01T00:00:00.000Z
          x-isnullable: true
      required:
        - title
        - author
        - message
    Threads:
      type: array
      items:
        $ref: '#/components/schemas/Thread'
...
    Post:
      description: |
        Сообщение внутри ветки обсуждения на форуме.
      type: object
      properties:
        id:
          type: number
          
          description: Идентификатор данного сообщения.
          readOnly: true
        parent:
          type: number
          
          description: |
            Идентификатор родительского сообщения (0 - корневое сообщение обсуждения).
        author:
          type: string
          format: identity
          description: Автор, написавший данное сообщение.
          example: j.sparrow
          x-isnullable: false
        message:
          type: string
          format: text
          description: Собственно сообщение форума.
          example: We should be afraid of the Kraken.
          x-isnullable: false
        isEdited:
          type: boolean
          description: Истина, если данное сообщение было изменено.
          readOnly: true
          x-isnullable: false
        forum:
          type: string
          format: identity
          description: Идентификатор форума (slug) данного сообещния.
          readOnly: true
        thread:
          type: number
          
          description: Идентификатор ветви (id) обсуждения данного сообещния.
          readOnly: true
        created:
          type: string
          format: date-time
          description: Дата создания сообщения на форуме.
          readOnly: true
          x-isnullable: true
      required:
        - author
        - message
...
    PostUpdate:
      description: |
        Сообщение для обновления сообщения внутри ветки на форуме.
        Пустые параметры остаются без изменений.
      type: object
      properties:
        message:
          type: string
          format: text
          description: Собственно сообщение форума.
          example: We should be afraid of the Kraken.
    PostFull:
      type: object
      description: |
        Полная информация о сообщении, включая связанные объекты.
      properties:
        post:
          $ref: '#/components/schemas/Post'
        author:
          $ref: '#/components/schemas/User'
        thread:
          $ref: '#/components/schemas/Thread'
        forum:
          $ref: '#/components/schemas/Forum'
...
    Slug:
      type: string
      format: identity
    Limit:
      type: number
      
      default: 100
      minimum: 1
      maximum: 10000
    Since:
      type: string
      format: date-time
    Desc:
      type: boolean
    Related:
      type: array
      items:
        type: string
        enum:
          - user
          - forum
          - thread
    Id:
      type: number
...